APM-1112 moved log history to the app level
