+++ projects/ama-sdk/src/lib/logging/components/log-history/log-history-entry/log-history-entry.component.ts
@@ -16,7 +16,7 @@
  */
 
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
-import { LogMessage } from '../../../interfaces';
+import { LogMessage } from '../../../../store/app.state';
 
 @Component({
     selector: 'amasdk-log-history-entry',
