+++ projects/ama-sdk/src/lib/logging/utils/logging.functions.ts
@@ -16,8 +16,9 @@
  */
 
 import { LogAction } from '../store/logging.actions';
-import { MESSAGE, LogMessageInitiator } from '../interfaces';
 import { arrayize } from '../../helpers/utils/arrayize';
+import { LogMessageInitiator } from '../../store/app.state';
+import { MESSAGE } from '../../store/app.state';
 
 export function logInfo(initiator: LogMessageInitiator, messages: string | string[]) {
     return new LogAction({
