+++ src/app/app/app-layout/app-layout.component.html
@@ -30,6 +30,7 @@
     <adf-sidenav-layout-content>
         <ng-template>
             <router-outlet></router-outlet>
+            <amasdk-editor-footer class="editor-footer"></amasdk-editor-footer>
         </ng-template>
     </adf-sidenav-layout-content>
