+++ src/app/process-editor/components/process-editor/process-editor.component.html
@@ -47,7 +47,6 @@
                     </div>
                 </mat-tab>
             </mat-tab-group>
-            <amasdk-editor-footer class="editor-footer"></amasdk-editor-footer>
         </div>
     </ng-container>
