+++ src/app/process-editor/components/process-modeler/process-modeler.component.ts
@@ -24,11 +24,11 @@ import {
     SelectModelerElementAction,
     ChangedProcessAction,
     RemoveDiagramElementAction,
-    ToolbarMessageAction
 } from '../../store/process-editor.actions';
 import { ProcessEntitiesState } from '../../store/process-entities.state';
 import { ProcessDiagramLoaderService } from '../../services/process-diagram-loader.service';
 import { createSelectedElement } from '../../store/process-editor.state';
+import { ToolbarMessageAction } from '../../../../app/store/actions/app.actions';
 
 @Component({
     selector: 'ama-process-modeler',
