+++ src/app/process-editor/store/process-editor.reducer.ts
@@ -27,15 +27,9 @@ import {
     UPDATE_PROCESS_SUCCESS,
     VALIDATE_PROCESS_ATTEMPT,
     UPDATE_PROCESS_ATTEMPT,
-    SetLogHistoryVisibilityAction,
-    SET_LOG_HISTORY_VISIBILITY,
-    CLEAR_LOG_HISTORY,
-    TOOLBAR_MESSAGE,
-    ToolbarMessageAction,
     UPDATE_PROCESS_FAILED
 } from './process-editor.actions';
-import { OPEN_CONFIRM_DIALOG, LOG_ACTION, LogAction } from 'ama-sdk';
-import { PROCESS_EDITOR_LOGS } from '../services/process-editor.constants';
+import { OPEN_CONFIRM_DIALOG } from 'ama-sdk';
 
 export function processEditorReducer(
     state: ProcessEditorState = getInitialProcessEditorState(),
@@ -58,36 +52,6 @@ export function processEditorReducer(
 
         case REMOVE_DIAGRAM_ELEMENT:
             return removeElement(state, <RemoveDiagramElementAction> action);
-
-        case TOOLBAR_MESSAGE:
-            return {
-                ...state,
-                toolbar: {
-                    ...state.toolbar,
-                    userMessage: (<ToolbarMessageAction>action).message
-                }
-            };
-
-        case LOG_ACTION:
-            return storeLog(state, <LogAction> action);
-
-        case CLEAR_LOG_HISTORY:
-            return {
-                ...state,
-                toolbar: {
-                    ...state.toolbar,
-                    logs: []
-                }
-            };
-
-        case SET_LOG_HISTORY_VISIBILITY:
-            return {
-                ...state,
-                toolbar: {
-                    ...state.toolbar,
-                    logHistoryVisible: (<SetLogHistoryVisibilityAction>action).visible
-                }
-            };
 
         case UPDATE_PROCESS_ATTEMPT:
         case VALIDATE_PROCESS_ATTEMPT:
@@ -123,23 +87,6 @@ export function processEditorReducer(
     }
 
     return newState;
-}
-
-function storeLog(state: ProcessEditorState, action: LogAction): ProcessEditorState {
-    if (action.log.initiator.key !== PROCESS_EDITOR_LOGS) {
-        return state;
-    }
-
-    return {
-        ...state,
-        toolbar: {
-            ...state.toolbar,
-            logs: [
-                action.log,
-                ...state.toolbar.logs
-            ]
-        }
-    };
 }
 
 function setSelectedElement(state: ProcessEditorState, action: SelectModelerElementAction): ProcessEditorState {
